requirements.txt
Pillow~=8.1.2
opencv-python~=4.5.1.48
numpy~=1.19.5
matplotlib~=3.4.1
tensorflow~=2.7.0
scipy~=1.6.3
scikit-image~=0.18.1
tqdm~=4.63.0